feat: removed all database stuff in i28, added new file for types
